Add test for predict_lpc_high_order
This is the same data as in the non_subset test sample. That sample has
been verified against the reference decoder, so the expected output here
must be the correct output. This is more of a test that will catch bugs
early, and pinpoint the cause of a difference, if I ever change the
high-order LPC prediction code, than a test that improves coverage -- we
already have the test sample for that.