Fix integer underflow when reading Vorbis comments
When skipping a comment, the total number of expected comments goes
down. But it should not fall below zero.

This issue was benign: in the case of wraparound, we would still read to
the end of the block as usual, and no further. The 'comments_len'
variable was only used after that to report a format error in case of a
mismatch between how many comments we read, and how many the header said
there would be. The preallocation for the comments vector has already
happened at that point, so this could not cause a large allocation.

This issue was caught on CI by LibFuzzer.