Zero buffers before reading metadata into them
As Youngsuk Kim pointed out, these functions pass uninitialized memory
to a potentially user-provided ReadBytes implementation. That
implementation should never read the buffer, only write to it, but there
is nothing in the type system that would prevent it from reading the
buffer. I think this is mostly a hypothetical problem, but as this
happens in metadata reading, which is not as performance-critical as the
audio decoding, we can zero the buffer and eliminate any room for doubt.

To measure the performance impact, I ran Musium [1] a few times with a
warm page cache. In scanning mode, it relies heavily on the metadata
reading code in Claxon, so it should be an excellent benchmark for this
use case. Raw measurements, done on my i7 laptop, with CPU scaling
governor set to powersave but no other measures to reduce variance,
indiate a significant increase of instructions executed (about 4.9%),
but there is no evidence that this affects running time of the program.
I think this may be due to the superscalar CPU being very good at
zeroing. I expect that this would not hold on e.g. a Raspberry Pi, with
less instruction-level parallelism, but I have not measured this yet.

[1]: https://github.com/ruuda/musium

version  instructions   seconds
before   2,479,923,342  0.370411717
before   2,490,487,917  0.374051726
before   2,480,699,027  0.373158450
before   2,493,521,988  0.348720195
before   2,487,302,191  0.382991778
before   2,504,259,708  0.341248452
before   2,460,315,103  0.406407899
before   2,483,216,336  0.354381391
before   2,481,989,479  0.403849040
after    2,617,232,354  0.335858157
after    2,604,118,685  0.393060263
after    2,601,597,068  0.392742210
after    2,596,655,560  0.414775199
after    2,610,027,727  0.345395543
after    2,625,111,174  0.338605271
after    2,596,597,744  0.342955707
after    2,608,620,910  0.378825927
after    2,598,315,383  0.390962604
after    2,596,818,449  0.408997508
after    2,617,298,735  0.324540958

	Wilcoxon rank sum exact test

data:  isn_before and isn_after
W = 0, p-value = 1.191e-05
alternative hypothesis: true location shift is not equal to 0

	Wilcoxon rank sum exact test

data:  secs_before and secs_after
W = 52, p-value = 0.882
alternative hypothesis: true location shift is not equal to 0