Convert try! to ?-operator
If I am to bump the minimal supported Rust version, we might finally opt
in to the ?-operator.

The translation was done mechanically with rustfmt, with the following
settings in rustfmt.toml:

    use_try_shorthand = true
    format_strings = true

Changes were then manually picked with "git add --patch" to only include
the try!-to-? conversions, and not other formatting changes, to prevent
polluting the diff as much as possible. In some places formatting
changes were more invasive, but that brings this closer to what rustfmt
would do anyway, so I think that is fine.

This should be a costmetic change only, not a functional change.