Delete old metadata module
It has been (temporarily) superseded by ::metadata2. I’m not really
happy about how that one is structured, and besides it does not compile
any more on recent Rust versions, so I think I am going to add a
metadata3 module, and finally rename that one back to metadata. But for
now, get rid of the old module.