Backport bugfixes from v0.4.3
I started working on this metadata refactor before v0.4.3. Then a few
things were fixed on master. I rebased my refactor on top of master, but
because most of the code was moved to different files -- twice now --
these fixes were lost. Apply them again.

The test samples for the bugs are still present, so the tests would have
caught this later, but the tests don’t compile at the moment.

Because the minimal supported Rust version was bumped, supporting
FusedIterator is now also possible.