Add LLM policy to the contribution guide
I made this policy update in the RCL repository first. It applies to
Claxon for the same reasons, below is the commit message I used in the
RCL repository:

Now that "vibe coding" is starting to become a thing, and there are some
tasks in RCL that presumably could be vibe-coded, like support for more
editors, I think it's a real risk that people start making pull requests
for code that was primarily written by an LLM and not by a human.

I don't use such tools myself, but based on what I've seen, this
technology is at a level where it is able to write a (mostly) working
solution, but it is not nearly at a level where it produces acceptable
code. Output is far too verbose, and even when humans are driving, if
it's so easy to generate more code, then instead of thinking "how can I
abstract this so I don't have to duplicate it", you just accept the
generated duplicates. Kind of the antithesis of RCL itself.

Aside from that, there is still the legal grey zone of whether LLMs
enable mass copyright infringement by repeating snippets verbatim from
other codebases without attribution. I don't want to argue about where
the boundary is, I want to stay far away from the boundary.

In short, I don't want any AI slop in this codebase, so it's better to
warn people ahead of time to prevent disappointment.